requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
APScheduler>=3.10.0
PyYAML>=6.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0
plotly>=5.15.0
python-telegram-bot>=20.0
aiohttp>=3.9.0
